serialize test

+++ src/common/bls12377/Fp.sol
@@ -70,6 +70,6 @@ library BLS12FP {
     function serialize(Bls12Fp memory self) internal pure returns (bytes memory) {
         uint256 a = ByteOrder.reverse256(self.a);
         uint256 b = ByteOrder.reverse256(self.b);
-        return abi.encodePacked(b, a);
+        return abi.encodePacked(b, uint192(a));
     }
 }
